Factor out the option logic and pattern extension.
A temporary container `alternation` is used to hold the newly parsed
token; if the token is followed by an option sign (?), a special `none`
value is added to the container also.

The helper function `expand_alternation` factors out the extension of
all the patterns for each alternative in `alternation`.

The `none` value, of type `none_t`, a monostate, can be stored in the
`alternation` container through use of `std::variant`; this is aliased
to `maybe_token_t`, as it can be interpreted as a crude mimicry of the
Maybe types of other languages.

This clears the way for an alternation operator (|) to be added.