add alternation support
Add support for alternation. Alternatives are added until the
termination of a "final alternative"; this is detected as a space when
`final_alternative` is true.

The helper function`expand_alternatives` then creates new patterns for
each alternative; in order to capture the final alternation, it must be
called after the line is parsed, as well as each time a new alternation
starts.

As before, without an option (?) or an alternation (|) symbol,
`alternation` will only have one element.

Examples:

```
PATTERNS
pattern1

PATTERN pattern1
A | B
A:|:B

LEXICON A
a

LEXICON B
b
```

produces

```
0	1	a	a	0.000000
0	1	b	b	0.000000
0	1	a	@0@	0.000000
0	1	@0@	b	0.000000
1	0.000000
```