need to reset sieve watchdog after token
The `just_sieved` variable serves as a watchdog to make sure that, if we
have just seen a sieve operator, we afterwards see a token. But after
seeing a token, it needs to be reset. This was first noticed expanding
an alternation at the end of a forward sieve:

```
PATTERNS
a > b|c

LEXICON a
a

LEXICON b
b

LEXICON c
c
```