Fix abrupt exit on certain empty patterns
Pattern filtering by tags can leave a pattern with zero entries; in some
instances, the compiled transducer (which has no final states) was
integrated into the output, instead of being discarded. This results in
lttoolbox calling exit(3) with a "empty set of final states" error.

Watch more carefully for calls to minimise() and insertTransducer() and
discard the partial transducer so that this doesn't happen.