Unset anti-recursion flag in buildPatternSingleLexicon
We use NULL as an anti-recursion sentinal when building a pattern in
single-lexicon mode. If we notice that `patternTransducers[tok]` is NULL,
it means we were already trying to compile the pattern `tok`, and need
to exit.

But this flag was never un-set once the pattern was successfully
compiled; the next time this pattern needs to be built-or-looked up, we
would fail as if it were a recursive reference.

Unset the flag before returning after successful complication. A test is
added to catch such an error in the future.