Usage: clarify the various transducer file formats
Many new users have asked questions about the file formats, e.g.
expecting to be able to `cat` the HFST output, or expecting that the
AT&T format will show full string-pairs; clarify which file formats are
being produced.

In the future, it would be nice to add links to descriptions of these
formats.