add | (or), ^ (xor) tag operations with tests
Introduce new tag operations `|` and `^` ("or" and "xor", respectively).
They can be used as a term in a tag filter expression as `^[x,y,z,...]`.

The semantics of `^` are slightly subtle; the following
```
X[^[a,b]]
```
evaluates to
```
X[a,-b] | X[-a, b]
```

The semantics of `|` are more obvious:
```
X[|[a,b]]
```
is the same as
```
X[a] | X[b]
```