Allow empty patterns due to filtering and add verbose mode
If a pattern had a combination of filters that matched no entries,
this would (sometimes) be treated as an error in non-flag mode. Now it
produces a warning in verbose mode (#17) and leaves that pattern line
without a final (to be removed in minimization). This also makes the
behavior match flag-mode, which did not error in such cases.

Closes #22
Closes #17
Closes #12